Emilce/permisos

+++ app/Http/Controllers/Rol/RolController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\Rol;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Rol\StoreRolRequest;
 use App\Http\Requests\Rol\UpdateRolRequest;
+use App\Models\Acceso;
 use App\Models\Rol;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use RealRashid\SweetAlert\Facades\Alert;
 
 class RolController extends Controller
@@ -18,7 +20,8 @@ class RolController extends Controller
      */
     public function index()
     {
-        $roles = Rol::all();
+        $this->authorize('view', new Rol());
+        $roles = Rol::orderBy('descripcion', 'asc')->get();
         return view('rol.index')->with('roles', $roles);
     }
 
@@ -29,7 +32,12 @@ class RolController extends Controller
      */
     public function create()
     {
-        return view('rol.create');
+        $this->authorize('create', new Rol());
+        $accesos = Acceso::where('nivel', '>', 1)
+            ->orderBy('modulo', 'asc')
+            ->orderBy('descripcion', 'asc')->get();
+        return view('rol.create')
+            ->with('accesos', $accesos);
     }
 
     /**
@@ -40,10 +48,39 @@ class RolController extends Controller
      */
     public function store(StoreRolRequest $request)
     {
-        $rol = new Rol($request->all());
-        $rol->save();
-        toast('Rol grabado correctamente','success');
-        return redirect()->route('rol.index');
+        $rol = new Rol();
+        $rol->nombre = $request->get('nombre');
+        $rol->descripcion = $request->get('descripcion');
+        $accesos = $request->input('acceso', []);
+        $crear = $request->input('crear', []);
+        $modificar = $request->input('modificar', []);
+        $eliminar = $request->input('eliminar', []);
+        $visualizar = $request->input('visualizar', []);
+        $imprimir = $request->input('imprimir', []);
+        $anular = $request->input('anular', []);
+        if ($accesos != 'null') {
+            DB::transaction(function () use ($rol, $accesos, $crear, $modificar, $eliminar, $visualizar, $imprimir, $anular) {
+                $rol->save();
+                if ($accesos != 'null') {
+                    foreach ($accesos as $i => $acceso) {
+                        $rol->accesos()->attach($rol, [
+                            'acceso_id' => $acceso,
+                            'crear' => in_array($acceso, $crear),
+                            'modificar' => in_array($acceso, $modificar),
+                            'eliminar' => in_array($acceso, $eliminar),
+                            'visualizar' => in_array($acceso, $visualizar),
+                            'imprimir' => in_array($acceso, $imprimir),
+                            'anular' => in_array($acceso, $anular),
+                        ]);
+                    }
+                }
+            });
+            toast('Rol grabado correctamente', 'success');
+            return redirect()->route('rol.index');
+        } else {
+            toast('Debe ingresar al menos un acceso para grabar el permiso', 'warning');
+            return redirect()->route('rol.index');
+        }
     }
 
     /**
@@ -65,7 +102,15 @@ class RolController extends Controller
      */
     public function edit(Rol $rol)
     {
-        return view('rol.edit')->with('rol',$rol);
+        $this->authorize('update', $rol);
+        $accesos = Acceso::where('nivel', '>', 1)
+            ->orderBy('modulo', 'asc')
+            ->orderBy('descripcion', 'asc')->get();
+        $permisos = $rol->accesos()->get();
+        return view('rol.edit')
+            ->with('rol', $rol)
+            ->with('accesos', $accesos)
+            ->with('permisos', $permisos);
     }
 
     /**
@@ -77,10 +122,39 @@ class RolController extends Controller
      */
     public function update(UpdateRolRequest $request, Rol $rol)
     {
-        $rol->fill($request->all());
-        $rol->save();
-        toast('Rol editado correctamente','success');
-    	return redirect()->route('rol.index');
+        $rol->nombre = $request->get('nombre');
+        $rol->descripcion = $request->get('descripcion');
+        $accesos = $request->input('acceso', []);
+        $crear = $request->input('crear', []);
+        $modificar = $request->input('modificar', []);
+        $eliminar = $request->input('eliminar', []);
+        $visualizar = $request->input('visualizar', []);
+        $imprimir = $request->input('imprimir', []);
+        $anular = $request->input('anular', []);
+        if ($accesos != 'null') {
+            DB::transaction(function () use ($rol, $accesos, $crear, $modificar, $eliminar, $visualizar, $imprimir, $anular) {
+                if ($accesos != 'null') {
+                    $rol->accesos()->detach();
+                    foreach ($accesos as $i => $acceso) {
+                        $rol->accesos()->attach($rol, [
+                            'acceso_id' => $acceso,
+                            'crear' => in_array($acceso, $crear),
+                            'modificar' => in_array($acceso, $modificar),
+                            'eliminar' => in_array($acceso, $eliminar),
+                            'visualizar' => in_array($acceso, $visualizar),
+                            'imprimir' => in_array($acceso, $imprimir),
+                            'anular' => in_array($acceso, $anular),
+                        ]);
+                    }
+                }
+                $rol->update();
+            });
+            toast('Rol actualizado correctamente', 'success');
+            return redirect()->route('rol.index');
+        } else {
+            toast('Debe ingresar al menos un acceso para grabar el permiso', 'warning');
+            return redirect()->route('rol.index');
+        }
     }
 
     /**
@@ -91,9 +165,14 @@ class RolController extends Controller
      */
     public function destroy(Request $request)
     {
+        $this->authorize('delete', new Rol());
         $rol = Rol::findOrFail($request->id);
+        $accesos = $rol->accesos()->get();
+        foreach ($accesos as $acceso) {
+            $acceso->permisos->delete();
+        }
         $rol->delete();
-        toast('Rol eliminado correctamente','success');
+        toast('Rol eliminado correctamente', 'success');
         return redirect()->route('rol.index');
     }
 }
